+++ src/domain-model/contracts/domain-state-model.js
@@ -1,0 +1,10 @@
+export class DomainStateModel {
+  constructor({ domainName, ownerService, entityName, states = [] }) {
+    this.domainName = domainName;
+    this.ownerService = ownerService;
+    this.entityName = entityName;
+    this.states = Object.freeze([...states]);
+
+    Object.freeze(this);
+  }
+}